requirements.txt
python-libtorrent==2.0.5
msal==1.20.0
requests==2.28.1
python-dotenv==0.21.0
schedule==1.1.0
tqdm==4.64.1 
